Install dependencies in a job (#1)
Previously installed dependencies in a session. To ensure backward compatibility with engines, we need to install dependencies with a job until a forthcoming release of CML, when we may revert to sessions for installation. This is related to the treatment of `kernel` parameter. The tradeoffs between the two are minimal: a session creates less noise in the project and .project-metadata.yaml, but both methods result in dependencies being installed, and both methods are used in other AMPs.

Also remove the now-deprecated kernel parameter from the worker API calls, and update to 2021.09 runtime.